reestructuración proyecto

+++ Lab1/components/led_delay/include/led_delay.h
@@ -1,3 +1,3 @@
 #include <stdio.h>
 
-void delay_s(uint32_t int_s);
+void delay_s(uint64_t int_s);
