+++ Lab1/components/led_delay/led_delay.c
@@ -3,6 +3,6 @@
 #include "led_delay.h"
 #include "esp_rom_sys.h"
 
-void delay_s(uint32_t int_s) {
-    esp_rom_delay_us ((uint32_t) int_s * 1000000);
+void delay_s(uint64_t int_s) {
+    esp_rom_delay_us ((uint64_t) int_s * 1000000);
 }
